refacor(AboutComponent): move to app module folder
