+++ src/app/main/about/about.component.ts
